create FitService
